+++ src/Libraries/SmartStore.Services/Common/IGenericAttributeService.cs
@@ -69,5 +69,16 @@ namespace SmartStore.Services.Common
         /// <param name="value">Value</param>
 		/// <param name="storeId">Store identifier; pass 0 if this attribute will be available for all stores</param>
 		void SaveAttribute<TPropType>(BaseEntity entity, string key, TPropType value, int storeId = 0);
-    }
+
+		/// <summary>
+		/// Save attribute value
+		/// </summary>
+		/// <typeparam name="TPropType">Property type</typeparam>
+		/// <param name="entityId">Entity identifier</param>
+		/// <param name="key">The key</param>
+		/// <param name="keyGroup">The key group</param>
+		/// <typeparam name="TPropType">Property type</typeparam>
+		/// <param name="storeId">Store identifier; pass 0 if this attribute will be available for all stores</param>
+		void SaveAttribute<TPropType>(int entityId, string key, string keyGroup, TPropType value, int storeId = 0);
+	}
 }